ratings/requirements.txt
Flask>=2.3
pymongo>=4.6
mysql-connector-python>=8.1
